requirements.txt
# ParentingBench Requirements

# Core dependencies
pyyaml>=6.0
pydantic>=2.0

# LLM providers
openai>=1.0.0
anthropic>=0.7.0

# Optional: for future enhancements
# litellm>=1.0.0  # Unified LLM interface
# langchain>=0.1.0  # LangChain integration

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
